+++ src/util.js
@@ -1,5 +1,13 @@
 require('console-polyfill');
 
+var browser = require('./browser');
+
+var RollbarJSON = null;
+
+function setupJSON(JSON) {
+  RollbarJSON = JSON;
+}
+
 var parseUriOptions = {
   strictMode: false,
     key: [
@@ -130,7 +138,77 @@ function uuid4() {
 }
 
 
+// Modified version of Object.create polyfill from:
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/create
+function objectCreate(prototype) {
+  if (typeof Object.create != 'function') {
+    return ((function(undefined) {
+      var Temp = function() {};
+      return function (prototype) {
+        if(prototype !== null && prototype !== Object(prototype)) {
+          throw TypeError('Argument must be an object, or null');
+        }
+        Temp.prototype = prototype || {};
+        var result = new Temp();
+        Temp.prototype = null;
+
+        // to imitate the case of Object.create(null)
+        if(prototype === null) {
+           result.__proto__ = null;
+        }
+        return result;
+      };
+    })())(prototype);
+  } else {
+    return Object.create(prototype);
+  }
+}
+
+// IE8 logs objects as [object Object].  This is a wrapper that makes it a bit
+// more convenient by logging the JSON of the object.  But only do that in IE8 and below
+// because other browsers are smarter and handle it properly.
+function formatArgsAsString() {
+  var args = [];
+  for (var i=0; i < arguments.length; i++) {
+    var arg = arguments[i];
+    if (typeof arg === 'object') {
+      arg = RollbarJSON.stringify(arg);
+      if (arg.length > 500)
+        arg = arg.substr(0,500)+'...';
+    } else if (typeof arg === 'undefined') {
+      arg = 'undefined';
+    }
+    args.push(arg);
+  }
+  return args.join(' ');
+}
+
+function consoleError() {
+  if (browser.ieVersion() <= 8) {
+    console.error(formatArgsAsString.apply(null, arguments));
+  } else {
+    console.error.apply(console, arguments);
+  }
+}
+
+function consoleInfo() {
+  if (browser.ieVersion() <= 8) {
+    console.info(formatArgsAsString.apply(null, arguments));
+  } else {
+    console.info.apply(console, arguments);
+  }
+}
+
+function consoleLog() {
+  if (browser.ieVersion() <= 8) {
+    console.log(formatArgsAsString.apply(null, arguments));
+  } else {
+    console.log.apply(console, arguments);
+  }
+}
+
 var Util = {
+  setupJSON: setupJSON,
   isType: isType,
   parseUri: parseUri,
   parseUriOptions: parseUriOptions,
@@ -138,7 +216,11 @@ var Util = {
   sanitizeUrl: sanitizeUrl,
   traverse: traverse,
   typeName: typeName,
-  uuid4: uuid4
+  uuid4: uuid4,
+  objectCreate: objectCreate,
+  consoleError: consoleError,
+  consoleInfo: consoleInfo,
+  consoleLog: consoleLog
 };