+++ test/notifier.test.js
@@ -483,13 +483,23 @@ describe('Notfier.unhandledRejection()', function() {
   });
 
   context('with a null value', function() {
-    it('should not enqueue a payload', function(done) {
+    it('should enqueue a payload with trace and other details', function(done) {
       var err = null;
       var promise = {};
 
       notifier.unhandledRejection(err, promise);
 
-      expect(enqueueSpy.called).to.equal(false);
+      expect(enqueueSpy.calledOnce).to.equal(true);
+      var args = enqueueSpy.getCall(0).args;
+
+      var payloadData = args[0].data;
+      expect(payloadData.level).to.equal('error');
+      var description = args[0].data.body.trace.exception.description;
+      var message = args[0].data.body.trace.exception.message;
+      var frames = args[0].data.body.trace.frames;
+      expect(description).to.equal('unhandled rejection was null or undefined!');
+      expect(message).to.equal('unhandled rejection was null or undefined!');
+      expect(frames.length).to.equal(1);
       done();
     });
   })
@@ -644,6 +654,14 @@ describe('Notifier.uncaughtError()', function() {
     expect(args[2][3]).to.equal(33);
     expect(args[2][4]).to.equal(21);
 
+    done();
+  });
+
+  it('should return the uuid of the error payload sent to the server', function(done) {
+    var notifier = new Notifier();
+    var obj = notifier.error('testing return value');
+
+    expect(obj.uuid).to.be.a('string');
     done();
   });
 
@@ -2281,37 +2299,37 @@ describe('Notifier._messageIsIgnored()', function() {
   });
 
   it('should return false when a message does not match any ignored message', function(){
-    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: Github is down!']);
+    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: GitHub is down!']);
     var payload = buildPayloadWithExceptionMessage('Exception: Not all llamas are ugly.');
     expect(notifier._messageIsIgnored(payload)).to.equal(false);
   });
 
   it('should return false when a message does not have a trace', function(){
-    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: Github is down!']);
+    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: GitHub is down!']);
     var payload = buildPayloadWithExceptionMessage('Exception: Not all llamas are ugly.');
     delete payload.data.body.trace;
     expect(notifier._messageIsIgnored(payload)).to.equal(false);
   });
 
   it('should return false when a message does not have a trace.exception', function(){
-    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: Github is down!']);
+    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: GitHub is down!']);
     var payload = buildPayloadWithExceptionMessage('Exception: Not all llamas are ugly.');
     delete payload.data.body.trace.exception;
     expect(notifier._messageIsIgnored(payload)).to.equal(false);
   });
 
   it('should return false when a message does not have a trace.exception.message', function(){
-    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: Github is down!']);
+    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: GitHub is down!']);
     var payload = buildPayloadWithExceptionMessage('Exception: Not all llamas are ugly.');
     delete payload.data.body.trace.exception.message;
     expect(notifier._messageIsIgnored(payload)).to.equal(false);
   });
 
   it('should return true when a message has a body.message.body which matches the ignored message', function(){
-    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: Github is down!']);
+    var notifier = buildNotifierWithIgnoredMessages(['Error: MySpace profile contains no animated gif.', 'Warning: GitHub is down!']);
     var payload = buildPayloadWithExceptionMessage('Exception: Not all llamas are ugly.');
     delete payload.data.body.trace;
-    payload.data.body = {message: {body: 'Warning: Github is down!'}};
+    payload.data.body = {message: {body: 'Warning: GitHub is down!'}};
     expect(notifier._messageIsIgnored(payload)).to.equal(true);
   });